Move the media picker callbacks into a reusable helper class
