+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ai/AddProductWithAIFragment.kt
@@ -5,9 +5,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.result.ActivityResult
-import androidx.activity.result.contract.ActivityResultContracts.StartActivityForResult
-import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.viewModels
@@ -27,12 +24,8 @@ import com.woocommerce.android.ui.products.ai.ProductNameSubViewModel.NavigateTo
 import com.woocommerce.android.ui.products.ai.ProductNameSubViewModel.ShowMediaLibraryDialog
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ShowSnackbar
+import com.woocommerce.android.widgets.MediaPickerHelper
 import dagger.hilt.android.AndroidEntryPoint
-import org.wordpress.android.mediapicker.MediaPickerConstants
-import org.wordpress.android.mediapicker.api.MediaPickerSetup
-import org.wordpress.android.mediapicker.api.MediaPickerSetup.DataSource
-import org.wordpress.android.mediapicker.api.MediaPickerSetup.DataSource.DEVICE
-import org.wordpress.android.mediapicker.ui.MediaPickerActivity
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -45,11 +38,7 @@ class AddProductWithAIFragment : BaseFragment() {
     lateinit var uiMessageResolver: UIMessageResolver
 
     @Inject
-    lateinit var mediaPickerSetupFactory: MediaPickerSetup.Factory
-
-    private val resultLauncher = registerForActivityResult(StartActivityForResult()) {
-        handleMediaPickerResult(it)
-    }
+    lateinit var mediaPickerHelper: MediaPickerHelper
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         return ComposeView(requireContext()).apply {
@@ -83,7 +72,7 @@ class AddProductWithAIFragment : BaseFragment() {
                 )
                 is ShowMediaLibraryDialog -> viewModel.onMediaLibraryDialogRequested()
 
-                is ShowMediaLibrary -> showMediaLibrary(event.source)
+                is ShowMediaLibrary -> mediaPickerHelper.showMediaPicker(event.source, ::showPackagePhotoBottomSheet)
 
                 is ShowSnackbar -> uiMessageResolver.showSnack(event.message)
 
@@ -92,32 +81,12 @@ class AddProductWithAIFragment : BaseFragment() {
         }
     }
 
-    private fun showPackagePhotoBottomSheet(imageUrl: String) {
-        findNavController().navigateSafely(
-            directions = AddProductWithAIFragmentDirections
-                .actionAddProductWithAIFragmentToPackagePhotoBottomSheetFragment(imageUrl)
-        )
-    }
-
-    private fun showMediaLibrary(source: DataSource) {
-        val mediaPickerIntent = MediaPickerActivity.buildIntent(
-            context = requireContext(),
-            mediaPickerSetupFactory.build(source)
-        )
-        resultLauncher.launch(mediaPickerIntent)
-    }
-
-    private fun handleMediaPickerResult(result: ActivityResult) {
-        if (result.resultCode == AppCompatActivity.RESULT_OK) {
-            result.data?.extras?.let { extra ->
-                val uri = (extra.getStringArray(MediaPickerConstants.EXTRA_MEDIA_URIS))
-                    ?.map { Uri.parse(it) }
-                    ?.first()
-
-                if (uri != null) {
-                    showPackagePhotoBottomSheet(uri.toString())
-                }
-            }
+    private fun showPackagePhotoBottomSheet(imageUri: Uri?) {
+        if (imageUri != null) {
+            findNavController().navigateSafely(
+                directions = AddProductWithAIFragmentDirections
+                    .actionAddProductWithAIFragmentToPackagePhotoBottomSheetFragment(imageUri.toString())
+            )
         }
     }
